# Internal: add push-slice-pop optimization