@sledorze 2018-02-23
@gcanti I think it's a breaking change as the reporting will immediately break for end user that have implemented custom Types with the validate API (public).

---

Also, if a 'no breaking change' is no more a constraint; what would we choose?

From my subjective perspective we have two solutions that look similar in speed gain but:
- One is mutable (brittle) and the user should know it.
- The other has a different report structure, (which I see as a feature - having more information; the hierarchy) but is a little more subtil in the parameters to pass to the validate function (the additional  key).

I'm wondering if we should not try decide first of a solution (with enhancement or maybe another alternative) before implementing one in a hurry and change it at the next breaking change (soon).

---

I am saying that as 2.8 is around the corner and Yes, I would love to have a nicer solution than intersecting with partials to have optional members.. :)