@gcanti 2018-02-24
> I think it's a breaking change as the reporting will immediately break for end user that have implemented custom Types with the validate API

You mean if the end user makes use of something else to create a failure other than the `failure` API right?

```ts
const MyString = new t.Type<string>(
  'string',
  (m): m is string => typeof m === 'string',
  // refuse to use the `failure` API here
  (m, c) => (typeof m === 'string' ? t.success(m) : t.failures([t.getValidationError(m, c)])),
  t.identity
)

const Person = t.type(
  {
    name: MyString,
    age: t.number
  },
  'Person'
)

console.log(PathReporter.report(Person.decode({})))
/*
[ 'Invalid value undefined supplied to : Person', // <= wrong error message here
  'Invalid value undefined supplied to : Person/age: number' ]
*/
```

That's my main concern too, if the `failure` API is always used then you won't notice a different behaviour. 

However I can't be sure of that in user land, so this change would be too dangerous.

Closing this, thanks for reviewing (and for the great work you did in the last few days on this subject).