@sledorze 2018-02-24
@gcanti ah.. that's a good question; it may break 'working' but wrong code, so I think yes, we should release as a bug fix.